+++ src/blade-runner.js
@@ -24,11 +24,12 @@ import * as Chat from '@system/chat';
 import * as BRMacro from '@system/macros';
 import BRRollHandler from '@components/roll/roller';
 import { registerSheets } from '@system/sheets';
-import { initializeHandlebars } from '@system/handlebars';
+import { initializeHandlebars, registerCustomItemTypesTemplates } from '@system/handlebars';
 import { changeEditorFont, registerSystemSettings } from '@system/settings';
 import { enrichTextEditors } from '@system/enricher';
 import { registerDiceSoNice } from './plugins/dice-so-nice';
 import itemPilesConfig from './plugins/item-piles';
+import { registerFontEditor } from './plugins/font-editor';
 // ! import { overrideInlineRollListener } from '@components/roll/inline-roll';
 import { getManual } from '@utils/get-manual';
 import BladeRunnerActor from '@actor/actor-document';
@@ -95,6 +96,7 @@ Hooks.once('init', async () => {
   registerSheets();
   registerSystemSettings();
   enrichTextEditors();
+  registerFontEditor();
   await initializeHandlebars();
 
   // Adds a shortcut directory for vehicle actors.
@@ -124,6 +126,8 @@ Hooks.once('ready', () => {
 
   // Replaces the a.inline listener with our own.
   // ! overrideInlineRollListener();
+
+  registerCustomItemTypesTemplates();
 
   console.log('Blade Runner RPG | Ready!');
   Hooks.callAll('bladeRunnerReady', game.bladerunner, CONFIG.BLADE_RUNNER);
@@ -207,7 +211,7 @@ Hooks.on('deleteActor', async actor => {
 
 Hooks.on('getChatMessageContextOptions', Chat.addChatMessageContextOptions);
 
-Hooks.on('renderChatLog', (_app, html, _data) => Chat.addChatListeners(html));
+Hooks.on('renderChatMessageHTML', (_msg, html, _data) => Chat.addChatListeners(html));
 Hooks.on('renderChatMessageHTML', (_msg, html, _data) => Chat.hideChatActionButtons(html));
 
 /* -------------------------------------------- */